(test_signaux) : tests signaux mais tjrs le pb que quand on connecte le dernier client l'autre s'arrête!!!

+++ server.py
@@ -7,7 +7,8 @@ import signal
 import sys
 from multiprocessing import Process, Manager
 from random import randint
-from process_player import *
+from threading import Semaphore
+
 
 #####INITIALISATION DICOS#####
 
@@ -70,37 +71,38 @@ def message_client(socket_player, message, retour="Nothing"):
             reponse = message_client(socket_player, message, retour)
     return reponse
 
-def stop_jeu(socket_player):
-    sig = signal.SIGUSR1
+def player_main(data, socket_player, que, sem_server, sem_player):
+    
+    message_client(socket_player, '0 Waiting')
     print(os.getpid(), "j'attends")
-    signal.sigwait([sig])
-    message_client(socket_player, '0 Fermeture')
-    print(os.getpid(), "je meurs")
-    time.sleep(5)
-    os.kill(os.getpid(), signal.SIGTERM)
-
-def player_main(data, socket_player, que):
-    on = True
-    thread = threading.Thread(target=stop_jeu, args=(socket_player, ))
-    thread.start()
-    while on:
-        message_client(socket_player, '0 Waiting')
-        print(data)
-        time.sleep(5)
-
-
+    
+    def handler(sig, frame):
+        if sig == signal.SIGUSR1 :
+            message_client(socket_player, '0 Le deck est vide')
+            os.kill(os.getpid(), signal.SIGKILL)
+        if sig == signal.SIGUSR2 :
+            message_client(socket_player, '0 Un 5 a été défaussé')
+            os.kill(os.getpid(), signal.SIGKILL)
+        if sig == signal.SIGINT :
+            message_client(socket_player, '0 Tous les fuze token ont été utilisés')
+            os.kill(os.getpid(), signal.SIGKILL)
+    
+    signal.signal(signal.SIGUSR1, handler)
+    signal.signal(signal.SIGUSR2, handler)
+    signal.signal(signal.SIGINT, handler)
 
 ######MAIN######
 
 if __name__ == '__main__':
     with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as server_socket:
         HOST = "localhost"
-        PORT = 6729
+        PORT = 6761
         key = 128
         nb_joueurs = 0
         child_processes = []
-        liste_couleurs = ["bleu", "rouge", "vert", "noir", "jaune", "orange", "violet"]
-
+        liste_couleurs = ["bleu", "rouge", "vert", "noir", "jaune", "orange", "violet", "rose"]
+        sem_server = Semaphore(0)
+        sem_player = Semaphore(1)
 
         server_socket.bind((HOST, PORT))
         server_socket.listen(1)
@@ -126,7 +128,7 @@ if __name__ == '__main__':
                 if key is not None:
                     que = sysv_ipc.MessageQueue(key, sysv_ipc.IPC_CREAT)
 
-                p = Process(target=player_main, args=(gros_dico, client_socket, que))
+                p = Process(target=player_main, args=(gros_dico, client_socket, que, sem_server, sem_player ))
                 p.start()
                 child_processes.append(p.pid)
                 message_client(client_socket, "0 Vous êtes le joueur 0")
@@ -134,16 +136,49 @@ if __name__ == '__main__':
                 for i in range(nb_joueurs - 1):
                     client_socket, address = server_socket.accept()
                     print("Connected to client: ", address)
-                    p = Process(target=player_main, args=(gros_dico, client_socket, que))
+                    p = Process(target=player_main, args=(gros_dico, client_socket, que, sem_server, sem_player))
                     p.start()
                     child_processes.append(p.pid)
                     message_client(client_socket, f"0 Vous êtes le joueur {i + 1}")
 
                 gros_dico["turn"] = 0
-
                 time.sleep(5)
                 print(child_processes)
-                for pid in child_processes:
-                    os.kill(pid, signal.SIGUSR1)
+
+                memory = gros_dico["fuse_token"]
+                on = True
+                while on :
+                    time.sleep(5)
+                    gros_dico["fuse_token"] = 0
+                    sem_player.acquire()
+                    print(child_processes)
+                    if gros_dico["fuse_token"] == 0 :
+                        if memory != 1 :
+                            for pid in child_processes:
+                                print(os.getpid(), "0 dico vide")
+                                print(pid)
+                                os.kill(pid, signal.SIGUSR2)
+                                time.sleep(5)
+                                on = False
+                                que.remove()
+                        else :
+                            for pid in child_processes:
+                                os.kill(pid, signal.SIGINT)
+                                on = False
+                                que.remove()
+                        
+                    if gros_dico["deck"] == 0 :
+                        for pid in child_processes:
+                            os.kill(pid, signal.SIGUSR1)
+                            on = False
+                            que.remove()
+                    
+                    else :
+                        print("je suis dans la boucle")
+                    
+                    memory = gros_dico["fuse_token"]
+                    sem_server.release()
+
+
                 while True :
                     pass
